Refactored fetch_video_data
